jacobian: calculate and Jacobian
jacobian.hs is standalone program that just calls printJacobian phi3
with it's argument and prints result.
Compile it with ghc --make jacobian.hs
jacobian.py is a simple SymPy program to get needed Taylor series
coefficients for the Jacobian. It needs number of loops to be passed as
an argument.
jacobian.sh is a simple shell script that calls jacobian first and that
jacobian.py. It needs two arguments: diagram signature and it's number
of loops (it's weird but let it be so for now).